Split port scanning and terminal helpers
